+++ src/comopnents/UsersList.tsx
@@ -1,11 +1,14 @@
 import classes from './User.module.css';
 import { UserItem } from './UserItem';
-import { User } from '../types/type';
+import { useContext } from 'react';
+import { UsersContext } from '../store/user-context';
 
-export const UsersList = ({ users }: { users: User[] }) => {
+export const UsersList = () => {
+	const usersCtx = useContext(UsersContext);
+
 	return (
 		<ul className={classes.lists}>
-			{users.map(item => {
+			{usersCtx.users.map(item => {
 				return <UserItem user={item} key={item.login.uuid} />;
 			})}
 		</ul>
